Endret feltet i backend/models til BinaryField da pickle returnerer bytes objekter. Endret jaccard_similarity til å ta inn to lister med lemmas, for raskere kode. La til en fil lemmalize som henter ut nøkkelordene til en setning (spørsmål) Crowbot_chat er endret for å tilpasses endringene i filene over.
