+++ crowbot/apiai_connection/crowbot_chat.py
@@ -1,12 +1,13 @@
 
 from django.conf import settings
 #settings.configure()
-import sys, json, codecs, apiai, re
+import sys, json, codecs, apiai, re, pickle
 import requests
 import django
 django.setup()
 from backend.models import Course, Question
 from backend.jaccard_similarity import *
+from backend.lemmalize import *
 
 
 
@@ -212,19 +213,23 @@ def ask_apiai(text):
             course = Course.objects.get(code=code)
             highest_ratio = 0
             similar_question = ''
+            lemmas1 = lemmalize(question)
+            lemmas_pickled = pickle.dumps(lemmas1)
             for q in Question.objects.all():
-                result = jaccard_similarity(question,q.text)
+                #ta inn lemma fra question
+                lemmas2 = pickle.loads(q.lemma)
+                result = jaccard_similarity(lemmas1,lemmas2)
                 if result[0]:
                     if result[1]>highest_ratio:
                         highest_ratio = result[1]
                         similar_question = q.text
             if similar_question == '':
-                Question.objects.create(text=question, course=course)
+                Question.objects.create(text=question, course=course, lemma=lemmas_pickled)
                 print(Question.objects.all())
                 return crowbot_answer(response)
             else:
                 print(Question.objects.all())
-                return ("Similar question detected:",similar_question)
+                return ("Similar question detected: {:s} with ratio {:.3}.".format(similar_question, highest_ratio))
                 #noe med at svaret til similar question presenteres for bruker
 
         except django.core.exceptions.ObjectDoesNotExist:
